Move container stats sal to tools
Fix syntax, refactor code to use tools and apply pre-commit
